FileNotFound
Error 404: Description not found.
This originally was a simple fix of @TheSwiftPhantom's FileNotFound Chomusuke tweak, but I decided to personalize it a bit. Lil-404 with rainbow eyes and no shadow stands on front of the iOS version of DDLC's "I gently open the door" error message. If you just want a fixed version of TSP's tweak, replace all "RenpyError.jpg" with "07.png" in public\styles\index.css file.